# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sehhat Foundation Online Exam System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /faculty/exams:
    get:
      operationId: getFacultyExams
      tags: [faculty]
      summary: List exams for faculty
      responses:
        "200":
          description: List of exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
    post:
      operationId: createExam
      tags: [faculty]
      summary: Create an exam slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExamRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
  /faculty/exams/{id}:
    get:
      operationId: getExamById
      tags: [faculty]
      summary: Get exam by id with MCQs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamWithMcqs"
    delete:
      operationId: deleteExam
      tags: [faculty]
      summary: Delete exam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /faculty/exams/{id}/mcqs/bulk:
    post:
      operationId: bulkUploadMcqs
      tags: [faculty]
      summary: Bulk upload MCQs for an exam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkMcqRequest"
      responses:
        "200":
          description: MCQs saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /faculty/mcqs/{id}:
    put:
      operationId: updateMcq
      tags: [faculty]
      summary: Update a single MCQ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/McqInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mcq"
    delete:
      operationId: deleteMcq
      tags: [faculty]
      summary: Delete a single MCQ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /student/list:
    get:
      operationId: getStudentList
      tags: [students]
      summary: Get list of students for a program/year
      parameters:
        - name: programId
          in: query
          required: true
          schema:
            type: integer
        - name: yearId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentInfo"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    StudentInfo:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        rollNumber:
          type: string
          nullable: true
      required:
        - id
        - name
...
    Exam:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        subjectId:
          type: integer
        subjectName:
          type: string
          nullable: true
        programId:
          type: integer
        programName:
          type: string
          nullable: true
        semesterId:
          type: integer
        yearId:
          type: integer
          nullable: true
        durationMinutes:
          type: integer
        totalQuestions:
          type: integer
        isActive:
          type: boolean
        mcqCount:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - subjectId
        - programId
        - semesterId
        - durationMinutes
        - totalQuestions
        - isActive
        - createdAt
    CreateExamRequest:
      type: object
      properties:
        title:
          type: string
        subjectId:
          type: integer
        programId:
          type: integer
        semesterId:
          type: integer
        yearId:
          type: integer
          nullable: true
        durationMinutes:
          type: integer
        totalQuestions:
          type: integer
        isActive:
          type: boolean
      required:
        - title
        - subjectId
        - programId
        - semesterId
        - durationMinutes
        - totalQuestions
        - isActive
    Mcq:
      type: object
      properties:
        id:
          type: integer
        examId:
          type: integer
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctOption:
          type: string
          enum: [A, B, C, D]
        questionNumber:
          type: integer
      required:
        - id
        - examId
        - questionText
        - optionA
        - optionB
        - optionC
        - optionD
        - correctOption
        - questionNumber
    McqInput:
      type: object
      properties:
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctOption:
          type: string
          enum: [A, B, C, D]
        questionNumber:
          type: integer
      required:
        - questionText
        - optionA
        - optionB
        - optionC
        - optionD
        - correctOption
        - questionNumber
    BulkMcqRequest:
      type: object
      properties:
        mcqs:
          type: array
          items:
            $ref: "#/components/schemas/McqInput"
      required:
        - mcqs
    ExamWithMcqs:
      type: object
      properties:
        exam:
          $ref: "#/components/schemas/Exam"
        mcqs:
          type: array
          items:
            $ref: "#/components/schemas/Mcq"
      required:
        - exam
        - mcqs
...