# lib/api-zod/src/generated/types/createExamRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sehhat Foundation Online Exam System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateExamRequest {
  title: string;
  subjectId: number;
  programId: number;
  semesterId: number;
  yearId?: number | null;
  durationMinutes: number;
  totalQuestions: number;
  isActive: boolean;
}